+++ lib/abi.ex
@@ -5,7 +5,7 @@ defmodule ABI do
   it to or from types that Solidity understands.
   """
 
-  alias ABI.Util
+  alias ABI.{FunctionSelector, Parser, TypeEncoder, TypeDecoder, Util}
 
   @doc """
   Encodes the given data into the function signature or tuple signature.
@@ -44,12 +44,17 @@ defmodule ABI do
       ...> |> Base.encode16(case: :lower)
       "b85d0bd200000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001"
   """
-  def encode(function_signature, data) when is_binary(function_signature) do
-    encode(ABI.Parser.parse!(function_signature), data)
+
+  def encode(function_signature, data, data_type \\ :input)
+
+  def encode(function_signature, data, data_type) when is_binary(function_signature) do
+    function_signature
+    |> Parser.parse!()
+    |> encode(data, data_type)
   end
 
-  def encode(%ABI.FunctionSelector{} = function_selector, data) do
-    ABI.TypeEncoder.encode(data, function_selector)
+  def encode(%FunctionSelector{} = function_selector, data, data_type) do
+    TypeEncoder.encode(data, function_selector, data_type)
   end
 
   @doc """
@@ -76,12 +81,17 @@ defmodule ABI do
       ...> |> ABI.decode("b85d0bd200000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001" |> Base.decode16!(case: :lower))
       [<<1::160>>, true]
   """
-  def decode(function_signature, data) when is_binary(function_signature) do
-    decode(ABI.Parser.parse!(function_signature), data)
+
+  def decode(function_signature, data, data_type \\ :input)
+
+  def decode(function_signature, data, data_type) when is_binary(function_signature) do
+    function_signature
+    |> Parser.parse!()
+    |> decode(data, data_type)
   end
 
-  def decode(%ABI.FunctionSelector{} = function_selector, data) do
-    ABI.TypeDecoder.decode(data, function_selector)
+  def decode(%FunctionSelector{} = function_selector, data, data_type) do
+    TypeDecoder.decode(data, function_selector, data_type)
   end
 
   @doc """
@@ -106,11 +116,11 @@ defmodule ABI do
       ...> |> ABI.find_and_decode("b85d0bd200000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000001" |> Base.decode16!(case: :lower))
       {%ABI.FunctionSelector{type: :function, function: "bark", input_names: ["at", "loudly"], method_id: <<184, 93, 11, 210>>, returns: [], types: [:address, :bool]}, [<<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1>>, true]}
   """
-  def find_and_decode(function_selectors, data) do
+  def find_and_decode(function_selectors, data, data_type \\ :input) do
     with {:ok, method_id, _rest} <- Util.split_method_id(data),
          {:ok, selector} when not is_nil(selector) <-
            Util.find_selector_by_method_id(function_selectors, method_id) do
-      {selector, decode(selector, data)}
+      {selector, decode(selector, data, data_type)}
     end
   end
 
@@ -181,11 +191,11 @@ defmodule ABI do
   def parse_specification(doc, opts \\ []) do
     if opts[:include_events?] do
       doc
-      |> Enum.map(&ABI.FunctionSelector.parse_specification_item/1)
+      |> Enum.map(&FunctionSelector.parse_specification_item/1)
       |> Enum.reject(&is_nil/1)
     else
       doc
-      |> Enum.map(&ABI.FunctionSelector.parse_specification_item/1)
+      |> Enum.map(&FunctionSelector.parse_specification_item/1)
       |> Enum.reject(&is_nil/1)
       |> Enum.reject(&(&1.type == :event))
     end